requirements.txt
agithub==2.2.2
appnope==0.1.2
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
decorator==4.4.2
Deprecated==1.2.11
idna==2.10
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.14
ptyprocess==0.7.0
Pygments==2.7.4
PyJWT==1.7.1
python-dotenv==0.15.0
requests==2.25.1
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5
wrapt==1.12.1
